EUWG-106 & 107 - settings support and export have been added.

+++ openchrom/plugins/net.openchrom.xxd.classifier.supplier.ratios.ui/src/net/openchrom/xxd/classifier/supplier/ratios/ui/internal/provider/time/TimeRatioTableComparator.java
@@ -60,7 +60,7 @@ public class TimeRatioTableComparator extends AbstractRecordTableComparator impl
 					sortOrder = timeRatio2.getName().compareTo(timeRatio1.getName());
 					break;
 				case 2:
-					sortOrder = Double.compare(timeRatio2.getExpectedRetentionTime(), timeRatio1.getExpectedRetentionTime());
+					sortOrder = Integer.compare(timeRatio2.getExpectedRetentionTime(), timeRatio1.getExpectedRetentionTime());
 					break;
 				case 3:
 					sortOrder = Double.compare(timeRatio2.getDeviationWarn(), timeRatio1.getDeviationWarn());
@@ -96,7 +96,7 @@ public class TimeRatioTableComparator extends AbstractRecordTableComparator impl
 					sortOrder = timeRatio2.getName().compareTo(timeRatio1.getName());
 					break;
 				case 2:
-					sortOrder = Double.compare(timeRatio2.getExpectedRetentionTime(), timeRatio1.getExpectedRetentionTime());
+					sortOrder = Integer.compare(timeRatio2.getExpectedRetentionTime(), timeRatio1.getExpectedRetentionTime());
 					break;
 				case 3:
 					sortOrder = Double.compare(timeRatio2.getDeviation(), timeRatio1.getDeviation());
